Add timestamp to comment and post notifications (#2565)
This commit adds timestamps to the comment and post notifications.
Instead of using `strftime` and ordinalizing the date I decided to use the
`time_ago_in_words` methods. I think that this looks better and is more
accurate when it comes to notifications recently received. The final say
is up to the main contributors though, if you would like me to change it
back let me know.