Remove commented specs (#5096) [deploy]
* Remove comment spec with no examples

* Delete commented out spec

This test could be refactored to a green state, but I'm not sure it's
worth the effort.

I started down that path, but it was time consuming. It might be more
effective to rewrite this test from scratch.

For the time being, since it has not been running, I think it should be
removed from the codebase.