Fix "current period" admin charts data examples (#17118)
* Disable current period specs

These are failing about half of the time (when the test runs east of
UTC, like Perth or Astana, the counts are incorrect, when the test
runs west of UTC like America/Chicago or Alaska, it passes).

* Cheat by setting clock to noon UTC

This aligns article published dates in ruby with those calculated by
postgresql. Copy the decision made in
https://github.com/forem/forem/pull/3518 for the analytics service.

* Remove testing comments

I had added titles to some of the articles I wanted excluded from the
result to confirm they were being included.