Fixing variant query specs (#17597)
* Fixing variant query specs

As I was working on adding a new variant, I discovered that not all of
our variants were being tested.  This had to do with a faulty assumption
regarding the scoping of the `variant` local variable.

The change ensures that we declare a subject once for each defined
variant within their own `describe` block.  Without that `describe`
block, you end up repurposing the same subject.

I found this because I made a deliberate error in the upcoming variant
and checked the tests.  They passed.

Related to forem/forem#17584

* Update spec/services/articles/feeds/variant_query_spec.rb

Co-authored-by: Julianna Tetreault <32834804+juliannatetreault@users.noreply.github.com>

* Update spec/services/articles/feeds/variant_query_spec.rb

Co-authored-by: Julianna Tetreault <32834804+juliannatetreault@users.noreply.github.com>

Co-authored-by: Julianna Tetreault <32834804+juliannatetreault@users.noreply.github.com>