Fix article destroy action (#21780)
* Fix article destroy action

The article destory action throws a 503 HTTP status code which
that the server is too busy to actually handle the DELETE request
at the moment.

While digging, I noticed that the article's notifications are
removed straight away and therefore this may be the cause of the
server business.

Also, the while the article is being destroyed, its mentions will
also be destroyed due to the
```
  has_many :mentions, as: :mentionable,
  inverse_of: :mentionable, dependent: :destroy
```
association but this association uses the `dependent: :destroy`
which means that callbacks will be triggered and there may take
some time as well.

* Update Comment model spec

Since the `has_many` association has changed slightly, the related
spec need to be updated accordingly.

---------

Co-authored-by: Ben Halpern <bendhalpern@gmail.com>